+++ src/interface/theme.ts
@@ -51,7 +51,7 @@ export interface IClassMap extends IClasses {
 	readonly focusInputWrapper?: string;
 	/** Classes for alt label text */
 	readonly altLabel?: string;
-	/** Float Label Placeholder if it is not floating yet */
+	/** Placeholder label if it is not floating yet, select "placeholder" if value not set */
 	readonly floatLabelPlaceholder?: string;
 	readonly invalidCheckboxWrapper?: string;
 	/** classes common to all tooltips */